% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfor_uncertainty_funs.R
\name{addStatUncToExpDt}
\alias{addStatUncToExpDt}
\title{Add statistical uncertainty to expDt}
\usage{
addStatUncToExpDt(expDt, statUncDt)
}
\arguments{
\item{expDt}{a datatable which specifies a selection of
data points in the EXFOR database}

\item{statUncDt}{a datatable with statistical uncertainties}
}
\value{
The modifed \code{expDt}, which is in principle not
        necessary because \code{expDt} is changed by reference.
}
\description{
Add statistical uncertainty to expDt
}
\note{
IMPORTANT: The datatable \code{expDt} given as input
                 argument is modified by this function!
}
