% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfor_uncertainty_funs.R
\name{getAutoUncertainty}
\alias{getAutoUncertainty}
\title{Automatically construct uncertainties for an EXFOR entry}
\usage{
getAutoUncertainty(subent, sysDt, statDt, dataref = NULL)
}
\arguments{
\item{subent}{the EXFOR subentry}

\item{sysDt}{a datatable as obtained by 
function \code{\link{getSystematicUncertainty}}}

\item{statDt}{a datatable as obtained by
function \code{\link{getStatisticalUncertainty}}}

\item{dataref}{reference cross sections used for relative uncertainties.
If \code{NULL} (default) the cross sections in the subentry
serve as the reference.}
}
\value{
a datatable of the same format as 
        \code{\link{getSystematicUncertainty}}
}
\description{
Retrieves the uncertainty information in an EXFOR entry 
and applies some rules to construct uncertainties based 
on the information in the EXFOR subentry.
For instance, inconsistent uncertainty information is
}
