% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfor_uncertainty_funs.R
\name{getAvailableUncertainties}
\alias{getAvailableUncertainties}
\title{Constructs Uncertainties for a list of EXFOR subentries}
\usage{
getAvailableUncertainties(expDt, subentList, dataref.col = NULL)
}
\arguments{
\item{expDt}{datatable which specifies a selection of data points in
the EXFOR database}

\item{subentList}{a list of subentries that must contain the subentries 
indexed by rows in \code{expDt}}

\item{dataref.col}{name of the column in \code{expDt} containing the 
reference cross sections used as the basis for
relative uncertainties}
}
\value{
A datatable with uncertainty specifications.
        Relevant columns are \code{DIDX, ERRTYPE, UNC, INFO}.
}
\description{
This function takes a datatable \code{expDt} which specifies
a selection of data points from different subentries
and returns a datataable with associated uncertainties.
}
\note{
The returned uncertainties do not necessarily reflect
      one-to-one those in the EXFOR subentry. If uncertainties
      in a subentry are inconsistent, e.g., ERR-T does not match
      with individual contributions, extra uncertainties are
      introduced.
}
