% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfor_uncertainty_funs.R
\name{getSystematicUncertainty}
\alias{getSystematicUncertainty}
\title{Extract systematic uncertainties from EXFOR subentry}
\usage{
getSystematicUncertainty(subent, dataref = NULL)
}
\arguments{
\item{subent}{an EXFOR subentry}

\item{dataref}{reference cross sections used for relative uncertainties.
if NULL (default) equals the cross sections in the subentyr.}
}
\value{
a datatable with the columns \code{ERRTYPE = (sys-rel, sys-abs)},
        \code{UNC} and \code{INFO}.
}
\description{
Extract systematic uncertainties from EXFOR subentry
}
