% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfor_uncertainty_funs.R
\name{hasValidUncertainties}
\alias{hasValidUncertainties}
\title{Check if EXFOR subentry has valid uncertainty info}
\usage{
hasValidUncertainties(subentList, quiet = FALSE)
}
\arguments{
\item{subentList}{a list with EXFOR subentries}

\item{quiet}{should error information be printed}
}
\value{
boolean vector indicating whether elements in \code{subentList}
        contain valid uncertainty information.
}
\description{
Check if EXFOR subentry has valid uncertainty info
}
