% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfor_uncertainty_funs.R
\name{splitUncDt}
\alias{splitUncDt}
\title{Split uncertainty datatable}
\usage{
splitUncDt(uncDt)
}
\arguments{
\item{uncDt}{a datatable with uncertainty information}
}
\value{
a list with two datatables for systematic and
        statistical uncertainties, respectively.
}
\description{
Splits the uncertainty information in a datatable
as obtained by \code{\link{getAvailableUncertainties}}
into two datatables: one with systematic components
and another one with statistical components.
}
